Refactor service imports to use __init__.py exports
- Add __init__.py files to all service packages with proper exports
- Update all imports to use consistent pattern: from speedofsound.services.X import Service
- Fix multiline import noqa comment placement in application.py
- All lint and format checks pass
