+++ speedofsound/services/configuration/__init__.py
@@ -1,0 +1,5 @@
+from speedofsound.services.configuration.configuration_service import (
+    ConfigurationService,
+)
+
+__all__ = ["ConfigurationService"]
