+++ speedofsound/services/control/__init__.py
@@ -1,0 +1,3 @@
+from speedofsound.services.control.control_service import ControlService
+
+__all__ = ["ControlService"]
