+++ speedofsound/services/extension/__init__.py
@@ -1,0 +1,3 @@
+from speedofsound.services.extension.extension_service import ExtensionService
+
+__all__ = ["ExtensionService"]
