+++ speedofsound/services/orchestrator/__init__.py
@@ -1,0 +1,3 @@
+from speedofsound.services.orchestrator.orchestrator_service import OrchestratorService
+
+__all__ = ["OrchestratorService"]
