+++ speedofsound/services/recorder/__init__.py
@@ -1,0 +1,3 @@
+from speedofsound.services.recorder.recorder_service import RecorderService
+
+__all__ = ["RecorderService"]
