+++ speedofsound/services/transcriber/__init__.py
@@ -1,0 +1,3 @@
+from speedofsound.services.transcriber.transcriber_service import TranscriberService
+
+__all__ = ["TranscriberService"]
