+++ speedofsound/services/transcriber/apis/nvidia_riva_transcriber.py
@@ -15,9 +15,7 @@ from speedofsound.models import (
     TranscriberResponse,
     TranscriberType,
 )
-from speedofsound.services.configuration.configuration_service import (
-    ConfigurationService,
-)
+from speedofsound.services.configuration import ConfigurationService
 from speedofsound.services.transcriber.apis.base_nvidia_transcriber import (
     BaseNvidiaTranscriber,
 )
