+++ speedofsound/services/typist/__init__.py
@@ -1,0 +1,3 @@
+from speedofsound.services.typist.typist_service import TypistService
+
+__all__ = ["TypistService"]
